Make JAILER_BASE_DIR dependend on execution root (#481)
Make JAILER_BASE_DIR dependend on execution root

it was hardcoded before on /var/lib/aleph so it wasn't respecting the proper user settings
and was failling if the user didn't have the correct permission for this folder and was producing warning

Pass the settings to MicroVM instead of importing it so this module can stay separate

+++ src/aleph/vm/hypervisors/firecracker/microvm.py
@@ -22,7 +22,6 @@ from .config import Drive, FirecrackerConfig
 logger = logging.getLogger(__name__)
 
 VSOCK_PATH = "/tmp/v.sock"
-JAILER_BASE_DIRECTORY = "/var/lib/aleph/vm/jailer"
 DEVICE_BASE_DIRECTORY = "/dev/mapper"
 
 
@@ -95,9 +94,9 @@ class MicroVM:
         return f"vm-{self.vm_id}"
 
     @property
-    def namespace_path(self):
+    def namespace_path(self) -> str:
         firecracker_bin_name = os.path.basename(self.firecracker_bin_path)
-        return f"{JAILER_BASE_DIRECTORY}/{firecracker_bin_name}/{self.vm_id}"
+        return str(self.jailer_base_directory / firecracker_bin_name / str(self.vm_id))
 
     @property
     def jailer_path(self):
@@ -121,12 +120,14 @@ class MicroVM:
         self,
         vm_id: int,
         firecracker_bin_path: Path,
+        jailer_base_directory: Path,
         use_jailer: bool = True,
         jailer_bin_path: Optional[Path] = None,
         init_timeout: float = 5.0,
     ):
         self.vm_id = vm_id
         self.use_jailer = use_jailer
+        self.jailer_base_directory = jailer_base_directory
         self.firecracker_bin_path = firecracker_bin_path
         self.jailer_bin_path = jailer_bin_path
         self.drives = []
@@ -237,7 +238,7 @@ class MicroVM:
             "--gid",
             gid,
             "--chroot-base-dir",
-            JAILER_BASE_DIRECTORY,
+            str(self.jailer_base_directory),
             "--",
             "--config-file",
             "/tmp/" + str(self.config_file_path.name),
@@ -492,7 +493,8 @@ class MicroVM:
         logger.debug("Removing files")
         if self.config_file_path:
             self.config_file_path.unlink(missing_ok=True)
-        system(f"rm -fr {self.namespace_path}")
+        if Path(self.namespace_path).exists():
+            system(f"rm -fr {self.namespace_path}")
 
     def __del__(self):
         try:
